+++ src/utils/env.ts
@@ -1,3 +1,5 @@
+import 'dotenv/config';
+
 export const DATABASE_URL = process.env.DATABASE_URL;
 export const PORT = process.env.PORT;
 export const JWT_SECRET = process.env.JWT_SECRET;