+++ test/history.test.ts
@@ -15,7 +15,7 @@ describe('GET /histories', () => {
     user.id = userId;
     const token = tokens[tokens.length - 1];
     const res = await getHistories(token, userId);
-    expect(res.body.error).toBe(false);
+    expect(res.data.error).toBe(false);
 
     expect(res.status).toBe(200);
   });
@@ -23,20 +23,22 @@ describe('GET /histories', () => {
     const token = tokens[tokens.length - 1];
     const userId = user.id;
     const res = await getHistories(token, userId);
-    expect(res.type).toBe('application/json');
+    // console.log('🚀 ~ it ~ res:', res);
+    expect(res.headers['content-type']).toContain('application/json');
   });
   it('should return a array of history', async () => {
     const token = tokens[tokens.length - 1];
     const userId = user.id;
-    const { body } = await getHistories(token, userId);
-    const histories = body.data as Array<HistoryResponse>;
+    const { data } = await getHistories(token, userId);
+    const histories = data.data as Array<HistoryResponse>;
     expect(typeof histories).toBe('object');
   });
   it('should return a error 401 if token is invalid', async () => {
     const token = 'jasfbaskjdb';
     const userId = '2';
     const res = await getHistories(token, userId);
-    expect(res.body.error).toBe(true);
+    // console.log('🚀 ~ it ~ res:', res);
+    expect(res.data.error).toBe(true);
     expect(res.status).toBe(401);
   });
 });
@@ -48,21 +50,21 @@ describe('DELETE /histories', () => {
   //   expect(res.body.error).toBe(false);
   //   expect(res.status).toBe(200);
   // });
-//   it('should return a response with json', async () => {
-//     const { body } = await signIn();
-//     const token = body.data.token;
-//     const res = await getHistories(token);
-//     expect(res.type).toBe('application/json');
-//   });
-//   //   it('should return a array of history', async () => {
-//   //     const token = await signIn();
-//   //     const { body } = await getHistories(token);
-//   //     expect(typeof body.data).toBe('array');
-//   //   });
-//   //   it('should return a error 401 if token is invalid', async () => {
-//   //     const token = 'jasfbaskjdb';
-//   //     const res = await getHistories(token);
-//   //     expect(res.body.error).toBe(true);
-//   //     expect(res.status).toBe(401);
-//   //   });
+  //   it('should return a response with json', async () => {
+  //     const { body } = await signIn();
+  //     const token = body.data.token;
+  //     const res = await getHistories(token);
+  //     expect(res.headers['content-type']).toBe('application/json');
+  //   });
+  //   //   it('should return a array of history', async () => {
+  //   //     const token = await signIn();
+  //   //     const { body } = await getHistories(token);
+  //   //     expect(typeof body.data).toBe('array');
+  //   //   });
+  //   //   it('should return a error 401 if token is invalid', async () => {
+  //   //     const token = 'jasfbaskjdb';
+  //   //     const res = await getHistories(token);
+  //   //     expect(res.body.error).toBe(true);
+  //   //     expect(res.status).toBe(401);
+  //   //   });
 });