+++ test/setupMocks.ts
@@ -1,3 +1,13 @@
-// process.env.TF_CPP_MIN_LOG_LEVEL = '3';
+// // import '@tensorflow/tfjs-node';
+// // jest.mock('@tensorflow/tfjs-node');
+// import * as tf from '@tensorflow/tfjs-node';
 
-jest.mock('@tensorflow/tfjs-node');
+// type JestGlobal = typeof globalThis & {
+//     tf: typeof tf;
+// }
+
+// const jestGlobal =  globalThis as JestGlobal;
+// // Inisialisasi TensorFlow di awal
+// jestGlobal.tf = tf;
+
+// console.log('TensorFlow engine initialized');